# openapi.json
{"openapi":"3.1.0","info":{"title":"Task Management API","description":"API for managing tasks with printing capabilities","version":"1.0.0"},"paths":{"/api/v1/auth/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","description":"OAuth2 compatible token login, get an access token for admin access.","operationId":"login_for_access_token_api_v1_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v1_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Login For Access Token Api V1 Auth Token Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks":{"get":{"tags":["tasks"],"summary":"Read Tasks","description":"Retrieve tasks.\n\nArgs:\n    skip: Number of records to skip\n    limit: Maximum number of records to return\n    include_archived: Whether to include archived tasks in the result\n    state: Optional state to filter tasks by (todo, in_progress, done, archived)\n    db: Database session","operationId":"read_tasks_api_v1_tasks_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"include_archived","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Archived"}},{"name":"state","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tasks by state","title":"State"},"description":"Filter tasks by state"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"},"title":"Response Read Tasks Api V1 Tasks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["tasks"],"summary":"Create New Task","description":"Create new task.","operationId":"create_new_task_api_v1_tasks_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/due/":{"get":{"tags":["tasks"],"summary":"Read Due Tasks","description":"Retrieve all tasks that are due within the next 24 hours.","operationId":"read_due_tasks_api_v1_tasks_due__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Task"},"type":"array","title":"Response Read Due Tasks Api V1 Tasks Due  Get"}}}}}}},"/api/v1/tasks/random/":{"get":{"tags":["tasks"],"summary":"Get Random Task","description":"Get a random task, prioritizing tasks that are:\n1. Not completed\n2. Due sooner\n3. Not yet started\n\nNote: Archived tasks are excluded.","operationId":"get_random_task_api_v1_tasks_random__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/api/v1/tasks/search/":{"get":{"tags":["tasks"],"summary":"Search Tasks","description":"Search tasks by title or description.\n\nArgs:\n    q: Search query\n    include_archived: Whether to include archived tasks in the result\n    db: Database session","operationId":"search_tasks_api_v1_tasks_search__get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Search query","title":"Q"},"description":"Search query"},{"name":"include_archived","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Archived"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"},"title":"Response Search Tasks Api V1 Tasks Search  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["tasks"],"summary":"Read Task","description":"Get task by ID.","operationId":"read_task_api_v1_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["tasks"],"summary":"Delete Task Endpoint","description":"Archive a task by ID.\n\nArgs:\n    task_id: ID of task to archive\n    db: Database session","operationId":"delete_task_endpoint_api_v1_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["tasks"],"summary":"Update Task Endpoint","description":"Update an existing task.\n\nArgs:\n    task_id: ID of task to update\n    task: Updated task data\n    db: Database session\n\nReturns:\n    Updated task\n\nRaises:\n    HTTPException: If task not found","operationId":"update_task_endpoint_api_v1_tasks__task_id__patch","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/start":{"post":{"tags":["tasks"],"summary":"Start Task","description":"Mark a task as in progress and set the started_at timestamp.","operationId":"start_task_api_v1_tasks__task_id__start_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/complete":{"post":{"tags":["tasks"],"summary":"Complete Task","description":"Mark a task as completed and set the completed_at timestamp.","operationId":"complete_task_api_v1_tasks__task_id__complete_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}/print":{"post":{"tags":["tasks"],"summary":"Print Task","description":"Print a task using the specified printer (defaults to PDF).","operationId":"print_task_api_v1_tasks__task_id__print_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}},{"name":"printer_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Type of printer to use","title":"Printer Type"},"description":"Type of printer to use"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/maintenance":{"post":{"tags":["tasks"],"summary":"Trigger Maintenance","description":"Manually trigger the task maintenance job.\nThis will process due tasks and clean up old ones.","operationId":"trigger_maintenance_api_v1_tasks_maintenance_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Trigger Maintenance Api V1 Tasks Maintenance Post"}}}}}}},"/api/v1/tasks/{task_id}/reset-to-todo":{"patch":{"tags":["tasks"],"summary":"Reset Task To Todo Endpoint","description":"Reset a task to todo state and clear its progress timestamps.","operationId":"reset_task_to_todo_endpoint_api_v1_tasks__task_id__reset_to_todo_patch","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/db/init":{"post":{"tags":["admin"],"summary":"Init Db","description":"Initialize database by creating all tables.\nRequires admin authentication.","operationId":"init_db_api_v1_admin_db_init_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Init Db Api V1 Admin Db Init Post"}}}}},"security":[{"OAuth2PasswordBearer":[]},{"APIKeyHeader":[]}]}},"/api/v1/admin/db/migrate":{"post":{"tags":["admin"],"summary":"Run Migrations","description":"Run all pending Alembic migrations.\nRequires admin authentication.","operationId":"run_migrations_api_v1_admin_db_migrate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Run Migrations Api V1 Admin Db Migrate Post"}}}}},"security":[{"OAuth2PasswordBearer":[]},{"APIKeyHeader":[]}]}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_api_v1_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v1_auth_token_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Task":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"state":{"type":"string","enum":["todo","in_progress","done","archived"],"title":"State","default":"todo"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date"},"reward":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reward"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"started_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Started At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["title","description","id"],"title":"Task"},"TaskCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"state":{"type":"string","enum":["todo","in_progress","done","archived"],"title":"State","default":"todo"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date"},"reward":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reward"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"},"started_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Started At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["title","description"],"title":"TaskCreate"},"TaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string","minLength":1},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string","minLength":1},{"type":"null"}],"title":"Description"},"state":{"anyOf":[{"type":"string","enum":["todo","in_progress","done","archived"]},{"type":"null"}],"title":"State"},"due_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Due Date"},"reward":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reward"}},"type":"object","title":"TaskUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/token"}}},"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-Admin-Key"}}}}